Add support for postgres

+++ building/models.py
@@ -86,7 +86,7 @@ class Tile(models.Model):
 
 
 class Address(models.Model):
-    node_id = models.IntegerField(unique=True, null=False, db_index=True)
+    node_id = models.BigIntegerField(unique=True, null=False, db_index=True)
     lat = models.FloatField(null=False, db_index=True)
     lon = models.FloatField(null=False, db_index=True)
     street = models.CharField(max_length=200)
@@ -203,7 +203,7 @@ class Company(models.Model):
 
 
 class Building(models.Model):
-    way_id = models.IntegerField(unique=True, null=False, db_index=True)
+    way_id = models.BigIntegerField(unique=True, null=False, db_index=True)
     osm_raw = models.JSONField()
     area = models.FloatField(db_index=True)  # in m^2
     length = models.FloatField()  # in m
